% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R, R/argsparsR-methods.R
\docType{class}
\name{argsparsR-class}
\alias{argsparsR}
\alias{argsparsR,argsparsR-class}
\alias{argsparsR-class}
\alias{initialize,argsparsR-method}
\title{argsparsR class definition.}
\usage{
\S4method{initialize}{argsparsR}(.Object)

argsparsR(args = NULL)
}
\arguments{
\item{.Object}{an argsparsR object.}

\item{args}{matrix of argument definitions, as defined in the `Details` section.}
}
\value{
a \code{argsparsR} object compiled with the parameter definitions given.
}
\description{
\code{argsparsR} creates an instance of a \code{argsparsR} object.
Takes the definition of the arguments as only parameter.

Create an instance of a \code{argsparsR} object. Takes the definition of
the arguments as only parameter.
}
\details{
The argument list is defined as a 5-column matrix, whose columns are:

1. name of the argument (string, no spaces)

2. long flag, of the kind \code{--argument} (string, no spaces)

3. short flag, of the kind \code{-a} (string, no spaces)

4. argument (basic R-)type (one among character, logical, integer, numeric)

5. default value

Providing at least one among the long and the short flags is mandatory,
both is optional. Remember anyway that you are filling in an R array,
so if you don't specify an item you have to set it to \code{''}.

The \code{nrow} parameter of \code{matrix} is the number of argument
definitions given.

The \code{byrow=TRUE} option in the matrix definition is also preferred,
as this allows a much clearer representation than the R default behaviour
of column-wise packing of the data.

The argument list is defined as a 5-column matrix, whose columns are:

1. name of the argument (string, no spaces)

2. long flag, of the kind \code{--argument} (string, no spaces)

3. short flag, of the kind \code{-a} (string, no spaces)

4. argument (basic R-)type (one among character, logical, integer, numeric)

5. default value

Providing at least one among the long and the short flags is mandatory,
both is optional. Remember anyway that you are filling in an R array,
so if you don't specify an item you have to set it to \code{''}.

The \code{nrow} parameter of \code{matrix} is the number of argument
definitions given.

The \code{byrow=TRUE} option in the matrix definition is also preferred,
as this allows a much clearer representation than the R default behaviour
of column-wise packing of the data.
}
\section{Slots}{

\describe{
\item{\code{args}}{matrix of argument definitions.}

\item{\code{no.args}}{number of argument definitions.}

\item{\code{args.names}}{vector containing the names of the arguments.}

\item{\code{long.flags}}{vector of flags in long form (e.g. \code{--flag}).
Positions correspond to the positions of \code{args.names}.}

\item{\code{short.flags}}{vector of flags in short form(e.g. \code{-f}). Positions
correspond to the positions of \code{args.names}.}

\item{\code{args.types}}{type of the arguments. Position correspond to the
positions in \code{args.names}.}

\item{\code{args.defaults}}{list of default values for the arguments. Positions correspond
to the positions in \code{args.names}.}

\item{\code{values}}{list of the actual values of the parameters, as from the command line.}
}}
\examples{
args <- argsparsR(matrix(c(
                  'name', '--name', '-n', 'character', 'me',
                  'age' , '--age' , ''  , 'integer'  ,  0  ,
                  'town', ''      , '-t', 'character', 'Brusaporco (TV)'
                  ),
                  nrow = 3,
                  byrow = TRUE
                  )
         )
args <- argsparsR(matrix(c(
                  'name', '--name', '-n', 'character', 'me',
                  'age' , '--age' , ''  , 'integer'  ,  0  ,
                  'town', ''      , '-t', 'character', 'Brusaporco (TV)'
                  ),
                  nrow = 3,
                  byrow = TRUE
                  )
         )
}

