% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/argsparsR-methods.R
\docType{methods}
\name{parsR}
\alias{parsR}
\alias{parsR,argsparsR}
\alias{parsR,argsparsR-method}
\title{parse command line arguments.}
\usage{
parsR(x)

\S4method{parsR}{argsparsR}(x)
}
\arguments{
\item{x}{a \code{argsparsR} object.}
}
\value{
a list containing:

* \code{params}, a vector containing the parameter names

* \code{values}, a vector containing the values for the corresponding parameter in
\code{params}. The value is the value provided through command line
if present, and the default value otherwise.
}
\description{
Parse command line arguments according to the definitions
provided to the constructor \code{argsparsR}.
}
\details{
Errors are thrown in case of mismatch between the nominal type of
the parameter and the actual type of the value given.
}

